+++ packages/plugins/Gitcoin/package.json
@@ -9,6 +9,7 @@
     }
   },
   "dependencies": {
+    "@masknet/flags": "workspace:^",
     "@masknet/icons": "workspace:^",
     "@masknet/plugin-infra": "workspace:^",
     "@masknet/shared": "workspace:^",
