+++ packages/plugins/Gitcoin/src/SNSAdaptor/contexts/Donate/DonateDialog.tsx
@@ -1,3 +1,8 @@
+import { BigNumber } from 'bignumber.js'
+import { type FC, memo, useCallback, useLayoutEffect, useMemo, useState } from 'react'
+import { useAsync } from 'react-use'
+import { Sniffings } from '@masknet/flags'
+import { Icons } from '@masknet/icons'
 import { useSNSAdaptorContext } from '@masknet/plugin-infra/content-script'
 import {
     ChainBoundary,
@@ -11,20 +16,16 @@ import {
     useSelectFungibleToken,
     WalletConnectedBoundary,
 } from '@masknet/shared'
-import { EMPTY_LIST, isTwitter, NetworkPluginID } from '@masknet/shared-base'
+import { EMPTY_LIST, NetworkPluginID } from '@masknet/shared-base'
 import { ActionButton, makeStyles, ShadowRootTooltip } from '@masknet/theme'
 import { useChainContext, useFungibleTokenBalance, useWeb3Connection } from '@masknet/web3-hooks-base'
 import { formatBalance, type FungibleToken, rightShift, ZERO } from '@masknet/web3-shared-base'
 import { type ChainId, isNativeTokenAddress, SchemaType, useGitcoinConstants } from '@masknet/web3-shared-evm'
 import { Box, DialogActions, DialogContent, Typography } from '@mui/material'
-import { type FC, memo, useCallback, useLayoutEffect, useMemo, useState } from 'react'
-import { useAsync } from 'react-use'
 import { useDonateCallback } from '../../hooks/useDonateCallback.js'
 import { useI18N } from '../../../locales/i18n_generated.js'
 import type { GitcoinGrant } from '../../../apis/index.js'
-import { Icons } from '@masknet/icons'
 import { GiveBackSelect } from './GiveBackSelect.js'
-import { BigNumber } from 'bignumber.js'
 import { useShowResult } from '../ResultModal/index.js'
 import { getSupportedChainIds } from '../../../utils.js'
 
@@ -136,7 +137,7 @@ export const DonateDialog: FC<DonateDialogProps> = memo(({ onSubmit, grant, ...r
     const donate = useCallback(async () => {
         const hash = await donateCallback()
         if (typeof hash !== 'string') return
-        const cashTag = isTwitter() ? '$' : ''
+        const cashTag = Sniffings.is_twitter_page ? '$' : ''
         if (!token) return
         const uiAmount = formatBalance(amount.plus(tipAmount), token.decimals)
         const shareText = t.share_text({
