+++ packages/shared-base/src/NextID/index.ts
@@ -1,6 +1,6 @@
 import { ProfileIdentifier } from '@masknet/base'
-import { EnhanceableSite } from '../Site/index.js'
-import { NextIDPlatform, type BindingProof } from '../NextID/type.js'
+import { EnhanceableSite } from '../Site/types.js'
+import { NextIDPlatform, type BindingProof } from './types.js'
 import { createLookupTableResolver } from '../utils/index.js'
 
 export const resolveNextIDPlatformToNetwork = createLookupTableResolver<NextIDPlatform, EnhanceableSite | undefined>(
