+++ packages/shared-base/src/NextID/types.ts
