+++ packages/shared-base/src/Site/index.ts
@@ -1,20 +1,5 @@
 import { Sniffings } from '@masknet/flags'
-
-export enum EnhanceableSite {
-    Localhost = 'localhost',
-    Twitter = 'twitter.com',
-    Facebook = 'facebook.com',
-    Minds = 'minds.com',
-    Instagram = 'instagram.com',
-    OpenSea = 'opensea.io',
-    Mirror = 'mirror.xyz',
-}
-
-export enum ExtensionSite {
-    Dashboard = 'dashboard',
-    Popup = 'popup',
-    PopupConnect = 'popup-connect',
-}
+import { ExtensionSite, EnhanceableSite } from './types.js'
 
 const matchEnhanceableSiteHost: Record<EnhanceableSite, RegExp> = {
     [EnhanceableSite.Localhost]: /localhost/i,
