+++ packages/shared-base/src/Site/types.ts
@@ -1,0 +1,15 @@
+export enum EnhanceableSite {
+    Localhost = 'localhost',
+    Twitter = 'twitter.com',
+    Facebook = 'facebook.com',
+    Minds = 'minds.com',
+    Instagram = 'instagram.com',
+    OpenSea = 'opensea.io',
+    Mirror = 'mirror.xyz',
+}
+
+export enum ExtensionSite {
+    Dashboard = 'dashboard',
+    Popup = 'popup',
+    PopupConnect = 'popup-connect',
+}
