+++ packages/shared-base/src/index.ts
@@ -6,13 +6,14 @@ export * from './Messages/Mask.js'
 export * from './Messages/Shared.js'
 export * from './NameService/index.js'
 export * from './NextID/index.js'
-export * from './NextID/type.js'
-export * from './Persona/type.js'
+export * from './NextID/types.js'
+export * from './Persona/types.js'
 export * from './Plugin/index.js'
 export * from './Post/type.js'
 export * from './Pure/index.js'
 export * from './Routes/index.js'
 export * from './Site/index.js'
+export * from './Site/types.js'
 export * from './MIME/index.js'
 export * from './Subscriptions/index.js'
 export * from './Social/index.js'
